Structure refactoring changes
