+++ src/skibidi_orm/migration_engine/converters/sql_converter.py
@@ -1,7 +1,9 @@
 from abc import abstractmethod, ABC
 from skibidi_orm.migration_engine.operations.column_operations import ColumnOperation
 from skibidi_orm.migration_engine.operations.table_operations import TableOperation
-from skibidi_orm.migration_engine.operations.constraints import Constraint
+from skibidi_orm.migration_engine.adapters.database_objects.constraints import (
+    Constraint,
+)
 
 
 class SQLConverter(ABC):
