+++ src/skibidi_orm/migration_engine/operations/column_operations.py
@@ -1,8 +1,10 @@
 from __future__ import annotations
 from abc import ABC, abstractmethod
 from skibidi_orm.migration_engine.operations.operation_type import OperationType
-from skibidi_orm.migration_engine.operations.constraints import Constraint
-from skibidi_orm.exceptions.operations import IrreversibleOperationError
+from skibidi_orm.migration_engine.adapters.database_objects.constraints import (
+    Constraint,
+)
+from skibidi_orm.exceptions.operations_exceptions import IrreversibleOperationError
 from skibidi_orm.migration_engine.adapters.base_adapter import BaseTable, BaseColumn
 from typing import Any
 from dataclasses import dataclass, field
