+++ tests/migration_engine/converter/sqlite3/test_constraint_conversion.py
@@ -1,7 +1,7 @@
 from pytest import raises
-from skibidi_orm.exceptions.constraints import UnsupportedConstraintError
+from skibidi_orm.exceptions.constraints_exceptions import UnsupportedConstraintError
 from skibidi_orm.migration_engine.converters.sqlite3_converter import SQLite3Converter
-from skibidi_orm.migration_engine.operations.constraints import (
+from skibidi_orm.migration_engine.adapters.database_objects.constraints import (
     PrimaryKeyConstraint,
     UniqueConstraint,
     ForeignKeyConstraint,
